requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
joblib==1.3.2
multidict==6.0.4
numpy==1.26.3
profanity-check==1.0.3
profanityfilter==2.0.6
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
threadpoolctl==3.2.0
urllib3==2.1.0
yarl==1.9.4
